Updated snapshots for the 'ys_deploy' module and prefix handler.
