+++ .vortex/installer/tests/Fixtures/handler_process/_baseline/web/modules/custom/sw_deploy/sw_deploy.info.yml
