+++ .vortex/installer/tests/Fixtures/handler_process/hosting_acquia/docroot/modules/custom/sw_deploy/sw_deploy.info.yml
