+++ .vortex/installer/tests/Fixtures/handler_process/names/web/modules/custom/the_force_deploy/the_force_deploy.info.yml
@@ -1,0 +1,5 @@
+name: New hope Deploy
+type: module
+description: Repeatable deploy hooks for New hope site.
+core_version_requirement: ^11
+package: the_new_hope
