+++ .vortex/installer/tests/Fixtures/handler_process/profile_the_empire/web/profiles/custom/the_empire/the_empire.info.yml
@@ -29,7 +29,7 @@ dependencies:
   - toolbar
   - views_ui
   - sw_base
-  - ys_deploy
+  - sw_deploy
 
 themes:
   - seven
